PY-41923 Mark EPs for custom target expression and class stubs as dynamic
Also, mark them as @ApiStatus.Internal instead of @Deprecated. These EP however
won't be safe to use as soon as we decide, e.g. to extract Django support into
a standalone plugin, because unloading it won't trigger rebuilding of indexes
in the current implementation leaving unused custom data in them.

GitOrigin-RevId: 3dbe4bee616c500409a088b825e3252615167069