DefaultJdkConfigurator is not component anymore
It's necessary to configure JDK before creating a new project on the new project wizard, so we just trigger this configuration in ProjectSdksModel#reset via ProjectJdkTable#preconfigure. A DefaultJdkConfigurator service now used to detect the best JDK on current system. We could omit this service, but Android Studio wants to disable this.

GitOrigin-RevId: a0e14ca9760eba0f0f30c5cd186683a9118377a0