vfs: added API to get nio Path from VirtualFile
Breaking the abstraction of VirtualFile and allow to get an underlying Path
from a VirtualFile (where possible). There are already places in the code
where we do leak that abstraction (e.g. Storage subsystem).  Now all tricky
conversions between VirtualFile and File can be replaced with a clean new API.

It is now implemented only for LocalFileSystemBase. Next it would allow us to
deal with non-default NIO Filesystems

GitOrigin-RevId: 3b8eb6f2dfaa77a3742bb17643b304f74bfc8e7b