Fix long-broken logging in Bazel tests
Apparently for a while, warning-level messages have been silently
dropped when running tests in Bazel. This is because (1) we did not
add test-log.xml to our test data deps, and (2) TestLogger has
unfortunate default behavior when test-log.xml is missing
(I filed IDEA-239980 for that).

This change should make some Bazel test failures significantly easier
to diagnose. For example, when there are errors in PluginManagerCore,
we will finally be able to see the cause in the logs.

Additionally, we should no longer see this message in our target logs:
    "log4j:WARN No appenders could be found for logger [...]"

Bug: see IDEA-239980
Test: regression test added in //tools/adt/idea
Change-Id: I1bc8743adf8d17088b050fb5ad519070104340c7