DumbModeTask - update documentation, deprecate, force getEquivalenceObject to be used
We deprecate default constructor of the object to enforce every task to
implement the getEquivalenceObject() clearly. For example to avoid IDEA-240591
scenario from happening.

It is too easy to mistakenly pass a too generic object as the EquivalenceObject
and have two different tasks merged. To overcome that,
we change the implementation to return a pair a given EquivalenceObject and getClass()

GitOrigin-RevId: 73e5716a1296243a5077647fd6982531fe3e33d5