PY-41923 Remove non-dynamic and now obsolete PyTestLegacyInteropInitializer
Originally, it was intended to aid migration to the new pytest run configuration
format, but it turned out that about year and a half ago it stopped doing anything,
having been accidentally subscribed to a wrong message bus (application- instead
of the project-level one). Probably the fact that nobody noticed is a good
indication that it's no longer needed.

GitOrigin-RevId: 89a09f610fd465484ff7a9b426b585e39e96bb36