[editor-tabs] use presentable name as a key for UniqueNameBuilder
In the absence of other files with the same name, presentable name is returned for VirtualFileWrapper instances. So for two files with same name but different presentable names it does not make sense to generate unique path-based names. Which means that presentable name should be used as key for UniqueNameBuilder.

GitOrigin-RevId: 1079ed87962d76537bc1c5548df02225f56f4917