[ruby] Docker: try to get gem environment with 'bundle exec'
For Docker SDK retrieving 'gem env' is now made in 3 attempts.
The first one is to get it within standard `bundle exec`. If it fails, IDE tries the second attempt - get it via `bundle exec` with custom empty Gemfile. And the last attempt - is to get it without bundler.

RUBY-26392 Unable to run rspec tests when using a jruby-based docker image as a remote ruby interpreter

[IDEA-CR-62609]

GitOrigin-RevId: 528292971903d19514b783115039ee81907dc6ed