[vcs-log] fix top scrollpane border in changes browser when log is shown in editor
A UIUtil.KEEP_BORDER_SIDES property is required to be added to a scroll pane, so that the borders won't be removed by JBTabsImpl. Since a3f66efac9f2d1ee327de92aa5f48c99ffbb0120 centerPanel is not a scroll pane, so the property was added to the incorrect component.

GitOrigin-RevId: 18bcd13f39c1cdb4fc8604516312cab781f89383