dumb-service: use synchronization instead of EDT to process tasks
This change adds a side effect (we agreed with Peter Gromov to accept it).
Before: DumbModeTasks processing was stopped when Modality State was increased (e.g. a dialog shown)
After: Modality state is ignored when DumbModeTasks are processes

The change does not affect the transition to and from dumb mode (it still happens
in the EDT with the same (or less) modality state)

GitOrigin-RevId: e931690d11f5047ce265ead3f15833f5b8275867