[jps build] set 11 as minimum required version for JRE to run build process
Since we're going to migrate modules in IntelliJ project (including modules used in the build process) to Java 11 (IDEA-241075) we need to ensure that JRE which runs the build process can handle this. This may affect behavior in users' projects: before the change JDK 8 (9,10) configured in a project may be used to run the build process, now in such cases we'll run the build process under IDE's runtime, and fork javac process under JDK configured in the project.

GitOrigin-RevId: a13da7925e8cf633f675a0db9779346a3f8fac6a