Fix getConflictingInstanceof
The case when we are going up from or/and operand should be handled differently than when we are going down into if/loop/etc. condition. In the former case, we need to negate condition for or and stop at original expression. In the latter case, we should not negate.
Fixes IDEA-241354 Cast conflicts with instanceof false-positives
Fixes IDEA-217602 False positive 'cast conflicts with instanceof'
Fixes IDEA-212568 Cast conflict with instanceof false positive

GitOrigin-RevId: b0468a9f6e20ec19f00ed67ea099cc51b695c667