Bugfix: do not activate Run/Debug tool window when myActivateToolWindow is false (GO-9369, IDEA-CR-1543)
Before this fix if you pass 'false' to 'withActivateToolWindow' the tool window is shown but without focus. With the fix it is not shown in that case. To show it without focus new 'withFocusToolWindow' method is added.

Along the way with this fix we avoid activating tool window twice in case of 'true' passed to 'withActivateToolWindow'.

GitOrigin-RevId: 3b7928ba535ebbb365d7735034897d88f87f569e