Bugfix: do not activate Run/Debug tool window when myActivateToolWindow is false (GO-9369, IDEA-CR-1543)
Before this fix if you pass 'false' to 'withActivateToolWindow' the tool window is shown but without focus. With the fix it is not shown in that case. To show it without focus new 'withFocusToolWindow' method is added.

Along the way with this fix we avoid activating tool window twice in case of 'true' passed to 'withActivateToolWindow'.

Cherry-picked from 3d0887f0831b454fb697964fd50c984d559003d4

GitOrigin-RevId: 1a56911dcac0f85133210d98e3aa2889cff097a6