Indexing: simplify implementation of IndexUpdateRunner, drop CachedFileContentQueue and siblings.
CachedFileContentQueue is a very tricky data structure with too many responsibilities and hard-to-test concurrent implementation. There were at least two bugs discovered (IDEA-241361, IDEA-238381) causing dramatic deadlocks.

Given recent rewrite of IndexUpdateRunner to use NonBlockingReadAction it has become possible to avoid the loop "if user is active, cancel file indexing and repeat again" and avoid pushing the file content to the queue on PCE. It was the trickiest responsibility of the queue, which is not needed anymore.

GitOrigin-RevId: 16dd5eff1d5337b737419a7272e8eca4c841c4fe