[project model] weaken requirements for OrderEntry created by not yet committed Module or Library
In the current project model ModifiableRootModel::addModuleOrderEntry/addLibraryEntry remember instance of passed Module/Library and return it even its instance isn't committed yet. But cases when this behavior is important (e.g. in Project Structure dialog) already provides special RootConfigurationAccessor instance, so it looks like it isn't needed to require this if the default RootConfigurationAccessor is used.

GitOrigin-RevId: bf96d2fa965ca1cf370d2e9993aaa88bb03aa2ed