Restore capitalization of Modal Tasks to "Title"
It was set to title in 90a9944cbba7ec8017c3acd8d4959e3c4c01d331: the modal task is displayed in a dialog and thus should have Title capitalization (https://jetbrains.design/intellij/controls/progress_bar/#in-dialog), as opposed to background tasks which are displayed in the status bar and should have Sentence capitalization.

Then it was accidentally changed to ProgressTitle in 5e45112b714812a4bbcff11fe28ff17642c70e30 along with other progress titles.

This commit restores the capitalization to "Title".

GitOrigin-RevId: 87a65ff9d28f7e1211500a7513429160a1cc8e61