PY-42700 Offer f-string-like completion after "{" in non-interpolated strings
The main idea here is that we also automatically convert the containing string
literal into a proper f-string upon selecting one of such variants.

Hopefully, it addresses the common problem of forgetting to add "f" prefix
before an interpolated (in mind) string literal in a more discoverable fashion
than a dedicated intention or a postfix template would.

GitOrigin-RevId: f59065abeb4a3b0cd7a0de42c9c06c9fad318b19