[workspace model] support loading and unloading plugins defining custom module root types
Since properties of custom source roots are stored in xml format in the workspace model entities it's enough to just drop caches if type is changed. However it's also required to unregister plugin which provides root type after root types are changed to 'unknown' (otherwise when ModifiableRootModel is created it'll already use UnknownSourceRootType for these roots), and collect affected roots before creating ModifiableRootModel when a new root type is registered.

GitOrigin-RevId: b39e769a1a42b5119d3b7fff61156f7e9bf19882