[platform] IDEA-220863 Jump outside closing bracket for
This patch adds a new method to TabOutScopesTracker#registerScopeRange that registers a range as a tab-out scope. Since the existing method TabOutScopesTrackerImpl#registerEmptyScope is a special case of this new method with
an empty range its body was moved to the new method and refactored to support a range
and registerEmptyScope simply delegates its execution to the new method.

The current implementation correctly handles cases where the JAVA language is injected
into an injection point and there is a test that verifies it.

Signed-off-by: Nikita Eshkeev <nikita.eshkeev@jetbrains.com>

GitOrigin-RevId: e11a5658b674c0f39bbaada2f23ed16587b7d837