[devkit] fix "Update IDE from Sources" action: add proper groovy version to classpath
The action was broken after updating to Groovy 2.5 because paths to libraries were hardcoded. Now they are taken from the project configuration, and groovy-cli-commons is added to classpath (it contains groovy.util.CliBuilder which is used by Gant). (Dependency on groovy-cli-commons in intellij.idea.ultimate.build should be removed after we update jps-build-script-dependencies-bootstrap library.)

GitOrigin-RevId: 728ca2e1a00f920d4d8203436d57c984d8daf57d