PY-35921 Pick the content root containing *.iml as the base directory of the module
A module may have many content roots. We used to pick the first mentioned
in its settings. It didn't match the expectations. If you have a content
root that contains the *.iml file of the module, it is likely the one you
want to use.

If there are no content roots that contain your *.iml, then we still use
the first content root.

GitOrigin-RevId: 7670730eab25d0ff1e269cfec79f4d85a186f62e