PY-40974 Always treat environment variables as raw bytes
It's unnecessary and sometimes wrong to assume any encoding
for an environment variable. They can be encoded differently
depending on the requirements of programs that utilize them.
In our case we just need to pass all the environment
variables "as is" to the IDE process.

I've modified printenv.py so that it's the same version of
the script for both Linux and macOS and for all the Python
versions from 2.3 to at least 3.9.

GitOrigin-RevId: 40d73ef6eb56da8609d75fbce8a4a04c52e317b4