[file-history] update diff preview later to get the correct selection
When GraphTableModel is assigned a new visible pack, it calls fireTableDataChanged to notify data change. One of the listeners that get notified is JTable.tableChanged, which, among other things, clears selection, as it may be inconsistent with the new data. VcsLogGraphTable.updateDataPack restores selection after this happens.

Among other listeners, fireTableDataChanged triggers diff preview update. As AbstractTableModel.fireTableChanged iterates listeners last to first, diff preview update is called when selection is yet to be updated, so it ends up with a selection inconsistent with the new visible pack. This commit updates diff preview in an invokeLater section, so it could get the restored selection.

EA-220843

GitOrigin-RevId: 0de426d3ab50c27f3abd6cc27b3eabecb4bcab8b