configuration-store: store PER_OS settings in os-specific subfolders
This will allow settings synchronization plugins (such as Settings Repository and IDE Settings Sync) to store the settings in the default place. Storing settings in the default place will allow to easily fix numerous serious issues currently existing in sync plugins, which would be difficult to fix within the current layout (one of such issues, for example, is that if a sync plugin gets disabled, then settings are loaded from the default place, and these are old settings, not the ones the user was using within the settings sync plugin).

GitOrigin-RevId: a1c991f00887e4733ba861d2a888ba1eb41b42ab