[workspace model] Add VFS refresh in unit test mode at module instance creation
It fixes `ExecutionManagerTest`. We got SDK not found exception at JPS execution without direct
VFS refresh. How did this situation appear? The system throws FileNotFoundException at module
creation and `FileBasedStorage#processReadException`  marks this `ModuleFileStorage` as
`isBlockSavingTheContent = true`. Then deserializer reads SDK property for the module and
changes it to the actual for this execution environment and serializes the data back to the
file system. But because on blocked storage, changed SDK didn't reflect in the .iml file and
JPS tried to get the default SDK which didn't valid for this environment.

GitOrigin-RevId: 4975731886628d3ea752e996ad6ca36b6994b046