[jps build] don't show "Clean up old output directories" notification
Since IDEA-185574 is fixed (IDEA 2018.1) when you rename or remove a module, the obsolete output directory is deleted automatically when Build Project is invoked. This notification was needed to clean up old output directories which were left by IDEA 2017.3 or older versions. But now it isn't necessary and may be annoying: if you update the project from VCS and get a change where some modules were removed or renamed, and then reopen the project, you'll see the notification before Build Project cleans up the old directories automatically (IDEA-237388). Also marker file for the notification may be created in unexpected places (IDEA-202729). Removing the notification is the simplest way to fix these issues. We may assume that users of IntelliJ IDEA 2017.3 or older already took the opportunity to clean the old directories after updating to IDEA version between 2018.1 and 2020.1.

GitOrigin-RevId: 93d371f43f3b61e1e7d9040e44df0683f582e0c7