[vcs] Use more discoverable name for Annotate action
Given the negative connotation of the word "blame" we find it inappropriate to use in UI, and especially to switch to it from a more positive "annotate". To add, there are tools out there that use Annotate as well (Space, Atlassian products), and some of them are looking to switch. See IDEA-122424 for details.

However, using "annotate" as term is a major discoverability issue as Git and Svn use "blame" as default term for the same. To solve this, we will use Annotate with Blame as the action name when only Git and Svn are used in the project.

GitOrigin-RevId: 3e219e99a1a4d170f1bab61e9594bd2bd159fe51