[intentions] IDEA-243633 Add a consistency test for unattended intentions descriptions
This patch adds a consistency test that goes through the
intellij.idea.ultimate.tests.main classpath, looks for inspections'
descriptions that do not have registered corresponding intention classes
that are retrieved with IntentionManager. Not all the intentions are
available at the classpath (e.g. CLion or AppCode) so their inspections'
descriptions are removed from the examination.

The test also generates a bash command to remove unattended intentions
if the test is run on Linux or MacOS.

This patch removes all the unattended intentions' descriptions files
that were discovered previously.

Signed-off-by: Nikita Eshkeev <nikita.eshkeev@jetbrains.com>

GitOrigin-RevId: 35e8799d1d40b506597052e61727543340d69c4a