[workspace model] WM-T-57 Update facet xml config in the store before commit
Prior to this fix, we don't have actual state of facet config in the store because it was serialized
only once at `ModifiableFacetModelBridge#addFacet`. But for example in Kotlin plugin, facet configuration
changed before the model commit at maven project import `MavenProjectImportHandler#configureFacet` and
these changes weren't reflected in the store.

GitOrigin-RevId: ec1227b91f87b440f4a58336598ab4feda4f9c57