[intentions] IDEA-243676 IDEA-243633 Add a consistency test for unattended intentions descriptions
This patch clean up the test code, it uses IntentionManagerSettings to get the list of intentions as metadata and extracts descriptionsDirectoryName. It helps to handle the cases when the description directory is configured through plugin.xml. In addition to that the tests also looks extracts the family names from intentions as IDEA looks for directories with this name if it's unable to find the directory by the classname.

Signed-off-by: Nikita Eshkeev <nikita.eshkeev@jetbrains.com>
(cherry picked from commit d730f15b65d778310e8e56ba8432eba19563e3b3)

GitOrigin-RevId: 4b18b18fd219a96dd08ce7caa6d6e51c1d07bcf2