Include cidr testFramework in the distribution
In order for plugins to write test IJ ships its testFramework.
We were not including that in our distrubution, and not only we
need to expose that so our C++ plugin can be built against a
prebuilt, also it's the right thing to do for Android Studio
plugin developers.

It only adds 500kb to the distribution zip.

Test: existing
Bug: 156497102
Change-Id: I77eaa17328c71e3a8dafb30587ba6045c94845fa