IDEA-CR-63904: node: fire single rootsChanged event on project open with several node_modules folders (WEB-45705, WEB-45655)
Previously, when a new node_modules folder was encountered during "Scanning files to index" phase, the folder was marked as excluded and rootsChanged event was fired to rebuild RootIndex. This led to firing as many rootsChanged events on project open as many new node_modules folders the project had.

Now, when a new node_modules is encountered during "Scanning files to index" phase:
1) all new node_modules are found in the project (or under a particular folder depending on the reason of "Scanning files to index")
2) the found node_modules folders are excluded and a single rootsChanged event is fired
The "Scanning files to index" phase is still running in parallel with these steps, but once the rootsChanged event is fired, it will be cancelled  and a new "Scanning files to index" phase will be scheduled.

To found all new node_modules folders another VFS scan is performed. Since it happens only when the fist new node_modules is already found, it won't happen on subsequent project openings or for projects without node_modules folder(s). When finding all new node_modules, node_modules shouldn't be scanned inside, because some node_modules can be very large - for this purpose, `com.intellij.openapi.roots.ContentIteratorEx` is introduced.
Also, `com.intellij.openapi.roots.impl.ProjectFileScanner` is introduced, it allows to visit all content files on project open and/or on new file/folder creation. Previously, `com.intellij.openapi.roots.impl.FilePropertyPusher` was used for this purpose, but it wasn't designed for this.

GitOrigin-RevId: e33d53d4fc31ea35ead915a5de7ee5406a28ce86