GTDU: change priority to references instead of declarations (as was before d7fc1e250deed0fd3d843b1542eb4d9bca2b0f2a)
The priority was changed to the declarations as part of IDEA-206322 to try and fix IDEA-194817 and GO-6572. This caused
numerous other issues (mostly API misuse: many references implement PsiNamedElement while they shouldn't), and also
an issue when both declaration and reference are equally valid choice IDEA-206322. GTDU was originally GTD,
so preferring references means going back to the roots.

GitOrigin-RevId: a52dcabe449f5cfe33031ce7415a55084310a3d5