IDEA-CR-64176: EA-209911 - assert: AbstractFileViewProvider.registerAsCopy: really check for cycles
Setting an original file for a file with existing copies appears to be a valid scenario.

For example, a completion is invoked in an empty literal where there's no injection, but there is one in the completion copy. Completion contributor copies the copy itself for some reason. Then a character is inserted into the very original document, so that there's no injection there as well. Another completion reuses the injection from the copy but now sets an original file for it.

(cherry picked from commit 472f6b8d73929dc2c596822b21f136f8d1ed92e2)

GitOrigin-RevId: fcb7fb49ee6863a0995f8d583f36b1d11589ed87