[workspace model storage] properly merge new external mapping after adding diff to a builder
Previously if an external mapping isn't present in the diff it is removed from the builder. But it may happen that a library is added and committed after a modifiable root model was created, so we need merge changes from the diff. Mappings are now automatically remove if there are no entries in them.

GitOrigin-RevId: 40890939ff6e7c0230339f96bc50ed1a26d16540