Fix warning in ScriptConfigurationHighlightingTest
In case of exception from dependencies resolver
we do not threat the configurations as up-to-date,
so the reloading process will be started during test
despite updateScriptDependenciesSynchronously called at startup.
This will lead to the warning from checkHighlighting
because if script reports are different we rehighlight the script.
Note that test won't fail.

Original commit: b3a84ba081d14306fa2301c31fff2706ef2ab319

GitOrigin-RevId: 0c77452108f98460efb080ee8a0c38524d55efba