[Shell] [Formatter] Make dumb formatter return leaf element with empty children.
Dumb formatter blocks of shell plugin are not leaf and at the same time has empty children. It leads to situation, that formatting block formally is not covering whole text. Because of it injected elements of shell cannot be reformatted.

Using `isLeaf = true` solves this problem and hopefully not introduce any other

GitOrigin-RevId: 6abc014898381423442b53f2aeb01f8a1608f956