[vcs-log] allow to use popups from toolbar in auto-popup mode
When toolbar is too big for the available space, an auto-popup is shown to allow access to the actions that do not fit. This auto-popup is closed on mouse exit, unless there are other popups opened. The information about opened popups is taken from the ActionManager.

By default, only ActionPopupMenuImpl instances register themselves in the ActionManager, and popups used in Vcs Log do not. This commit adds a listener to the popups shown from Vcs Log toolbar to add them to ActionManager when shown and remove when closed.

IDEA-238289

(cherry picked from commit 588dc5195e83bf67036cac0f7e1f5c581e69e1c5)

GitOrigin-RevId: 7cf8bfbc25febc3f9ba5c85e5ff328378171e842