configuration-store: rename State.additionalExportFile to additionalExportDirectory
This property is intended to represent only a directory, all usages are directories, and the code relies on the fact that it is a directory.

There are a couple of external plugins which declare the property, but they do it without any real need. Therefore it is enough to deprecate the old name without providing any additional migration.

GitOrigin-RevId: ce8963c265bd032ba8e9b1081263b2debb4c6b46