[workspace model] do not save configuration files if nothing was changed
If the entities weren't changed after the project was loaded from the configuration files, it makes no sense to save them. This would just take time and may lead to unnecessary changes in configuration files if they weren't formatted properly. E.g. if there is an additional empty line at the end of iml file it'll be removed if the module configuration is saved; however it isn't necessary to fix formatting so it's better to perform save only if something was changed in the module configuration, this way we won't get unnecessary modified files in a local copy.

GitOrigin-RevId: 92e52a6720fc829f1a904fafc31d9761d8a963a2