[vcs-log-graph] use last insertion point instead of a random occurrence to improve performance
ListIntToIntMap allows to map commit indexes from the full list to the partial list, where some of the commits are not visible. Long index is the index in the full list, short -- in the partial list. To implement the mapping, long list is divided into blocks of equal sizes. For each block it is calculated, how many commits are visible in this block and in previous blocks.

For getting long commit index for the short one, we first need to find the last block where visible commits count is less than short index and iterate from this block. Since c7344b55055788718f8071a3761a71eb1d4303c5 it was done with the Arrays.binarySearch, which does not guarantee to return the last block, thus leading to unnecessary iterations. To fix the problem, a custom method is used.

IDEA-241581

(cherry picked from commit 994963931de5cd121f7f889ab55ba28d8261f372)

GitOrigin-RevId: 7295e31637da8df0534fc0adaba3ccfe2169bcec