Prohibit overriding delegateProgressChange
because it's very brittle: some questionable inheritors used to call delegating action on the other indicator which was in incorrect state to break everything (e.g. calling popState() on not yet pushed indicator). These places rewritten to addStateDelegate() instead, with the delegate crafted to relay all UI changes like setText() to the needed indicator. To fix IDEA-244911 ProgressIndicator push/pop state does not work with delegation

GitOrigin-RevId: 93604687daf3ba276b84f28e77fdb47eedf3ee56