Explicitly set product prefixes in dev build numbers of PyCharm IDEs
Without it under debug the product build, as returned by ApplicationInfo.getBuild(),
contains only the numeric part specified in build.txt, e.g. "203.SNAPSHOT".

In particular, this version is then used in requests to Marketplace leading
to most of the plugins being treated as incompatible with debug instances of
the IDEs. FUS and CWM also have problems communicating with IDEs without
prefixes.

This feature of ApplicationInfo.xml was originally introduced in
68c820f66c9b687afd3a524ace7f0fc921d04ab5, and all other IDEs already follow
this scheme.

GitOrigin-RevId: 8fab6013d40db8de733ab79444345378907ff418