IDEA-243804: disable symbol-based UrlReferences
What happens:
- Java literal "http://localhost:8080/hello";
- the priority: own references > external references > old references;
- UrlReference which spans over the whole literal content falls into "external" category;
- contributed references url segment references are "old";
- UrlReference is preferred over segment reference.

To fix the issue "old" references for Java literals might have been returned as "own" like it was done for
XmlAttributeValue references; but such solution is fragile as we see in this very commit, there will always be another
element which should change its implementation to return "old" references as "own".

We need another solution, so the migration is effectively reverted.

GitOrigin-RevId: 2d048dee153ec2b879958f0dd0d1afea46c0a0d9