[workspace model] IDEA-245360 Fix the place of modules mavenize
This fix is needed for the new project model. Prior to it, `IdeModifiableModelsProvider`
was creating modifiable models which didn't contains changes which `setMavenizedModules`
did (change `entitySource`). Afterward  `IdeModifiableModelsProvider.commit` overridden
`entitySource` to the old one which was before the Mavenize.
To avoid this situation `setMavenizedModules` moved after `IdeModifiableModelsProvider.commit`
and in such cases all imported entities get updated `entitySource`.

GitOrigin-RevId: ae3d3f7a898f92d0606d1e9be77def056426ef6d