Do not generate resources_en.jar in for the android plugin
IJ build has a mechanism where it removes a bunch
of resources from the jars into resources_en.
We don't intentionally use this, but the patterns
defined in:

https://android.googlesource.com/platform/tools/idea/+/refs/heads/studio-master-dev/platform/build-scripts/groovy/org/jetbrains/intellij/build/impl/DistributionJARsBuilder.groovy#56

pull ~10 files into resource_en unnecesarily.

I'm not planning to support this in the unbundled build,
so I want the build to start reflecting that structure now.

Test: Manual
Bug: 156497102
Change-Id: I05d44a70c97a6dbbd92e9fada501c6f352a43b6b