[workspace model] IDEA-245561 Add explicit mapping from ContentRoot to SourceRoot
Prior to this fix, the process of detecting source roots for the correspondent content
root was based on the nearest folder. But this approach was not fully correct for content
roots e.g during gradle file modification (for the more precise example check created test
in this commit).  To avoid such detection in the new project model direct link between
ContentRoot and SourceRoot was created.

GitOrigin-RevId: 35756ca7f0d8482dc0d5031431e0bfc24c852c85