vcs: ignore ModalityState in invokeAfterUpdate
It affects a singe edge-case and is ignored otherwise.
Remove it to ensure that callbacks are handled in a unified way and to avoid confusion.

Usage in 'handleStoppedQueue' was introduced after everything else, in 134a1db80e0637ee75d60b9003eb6a2f830685db
Usage in 'runOrInvokeLaterAboveProgress' is ignored: modality from ProgressIndicator is used instead.

GitOrigin-RevId: 2255958a2aa6508a1e89426c79003bc63aa71c3f