[vcs-log] do not trigger tool window content initialization when creating VcsLogTabsWatcher
Method ToolWindow.getContentManager can trigger tool window content initialization if it was not performed already. During this process, main log is initialized, which requires an instance of VcsLogTabsWatcher. As the first VcsLogTabsWatcher is still inside a constructor and not assigned to the field yet, another instance is created. This leads to a memory leak, since one of the instances can not be disposed properly.

In order to avoid this issue, this commit uses ToolWindow.addContentManagerListener method to add a listener lazily.

(cherry picked from commit 073bb9d0775fa6f0244c7ab39307cea475c40ce5)

GitOrigin-RevId: ca365e8a8d82c0d5934952f5047cf5ae291a38d0