[codeInsight] IDEA-170168 Show method quick doc on modifiers
It used to be impossible to invoke the quick doc popup on the modifiers
of method, field or class. This patch adds the possibility to invoke the
quick doc popup on such modifiers. It does it by extending the
JavaDocumentationProvider#getCustomDocumentationElement method checking
if the passed contextElement element is instance of PsiKeyword and if so
it checks if it's parent excluding PsiModifierList, PsiReferenceList and
PsiTypeElement is an instance of PsiJavaDocumentedElement which means
that the keyword is either in the modifier list or throws list of a
method or in the class or field's modifier list. Now the user can press
CTRL + Q almost everywhere on the method's signature and review the
method's javadoc.

The implemented approach is flexible enough to handle the new features
like sealed classes and records.

Signed-off-by: Nikita Eshkeev <nikita.eshkeev@jetbrains.com>

GitOrigin-RevId: 3cc79b890d584196b19c64251217dc3951ab2a7e