[groovy] IDEA-242437: Add initial types to the state of first instruction of the DFA
InitialTypeProvider used to be invoked on-demand at READ instructions. This is wrong, because there are cases where there is no read instructions at all, but initial type influences the flow. Example: like in `if (...) { a = x }`. Here type of `a` after branching statement is a LUB of typeof(x) and initial type of a, but this initial type is not queried anywhere

GitOrigin-RevId: dca0ad8b158cac38e10c9c21970c4250648e0fee