[vcs-log] split VcsLogFile into two implementations
SimpleVcsLogFile keeps reference to VcsLogPanel and becomes invalid when corresponding VcsLogUiEx is disposed. Thus this file can be opened only once.

GitCompareBranchesFile creates new instances of VcsLogUiEx when editors are opened. This way multiple editors could be opened for this file, including in split mode, and file can be reopened from recent files list (to be implemented later).

GitOrigin-RevId: df3e1eef18b88656e902a78a9cf24136546a4d48